Temporary pause button for pwdr
This is the code I added to the pwdr firmware so that I would be able to use a switch to pause the printer while printing temporarily and after I hold down (toggle) the switch.
Once I release the switch, the printer continues printing the work piece from where it left off. This allows one to make modifications in the middle of a printing job to their work piece.
Switch terminal on one side is on pin 5 and ground and the other is on 5v supply.